Sketch out class-based alternative to procedural job
